+++ Player.cpp
@@ -377,13 +377,13 @@ void Player::ProcessCommand(QString command) {
         if ( inv == nullptr ) return;
         QString kind, sub;
         comm_stream >> kind >> sub;
-        const quint8 kind_code = BlockManager::StringToKind(kind);
+        const int kind_code = BlockManager::StringToKind(kind);
         if ( kind_code == LAST_KIND ) {
             emit Notify(tr("%1 command: invalid kind!").arg(request));
             return;
         } // else:
-        const quint8 sub_code = sub.isEmpty() ?
-            static_cast<quint8>(STONE) : BlockManager::StringToSub(sub);
+        const int sub_code = sub.isEmpty() ?
+            static_cast<int>(STONE) : BlockManager::StringToSub(sub);
         if ( sub_code == LAST_SUB ) {
             emit Notify(tr("%1 command: invalid substance!").arg(request));
             return;
